remove files and rename the new one (#1020)
