+++ Cargo.toml
@@ -11,7 +11,7 @@ with_mimalloc = ["dep:mimalloc"]
 cairo_1_tests = []
 
 [workspace]
-members = ["cli", "fuzzer", "rpc_state_reader", "rpc_state_reader_sn_api"]
+members = ["cli", "fuzzer", "rpc_state_reader"]
 
 [workspace.dependencies]
 cairo-vm = { version = "0.8.5", features = ["cairo-1-hints"] }
