+++ rpc_state_reader/Cargo.toml
@@ -6,17 +6,24 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-starknet_in_rust = { path = "../", version = "0.3.1" }
 ureq = { version = "2.7.1", features = ["json"] }
-serde = { version = "1.0.152", features = ["derive"] }
+serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0", features = [
     "arbitrary_precision",
     "raw_value",
 ] }
-starknet_api = { workspace = true }
+starknet_api = "0.4.1"
 cairo-lang-starknet = { workspace = true }
+cairo-lang-utils = { workspace = true }
 starknet = { workspace = true }
 thiserror = { workspace = true }
+flate2 = "1.0.25"
 serde_with = "3.0.0"
 dotenv = "0.15.0"
 cairo-vm = "0.8.5"
+blockifier = "0.2.0-rc0"
+starknet_in_rust = { path = "../", version = "0.3.1" }
+
+[dev-dependencies]
+pretty_assertions_sorted = "1.2.3"
+test-case = "3.1.0"
