+++ rpc_state_reader/src/utils.rs
